Moved posts.js from lib to utils since utils is used for functions folder
